Did some code restructure.
