+++ app/src/main/res/layouts/chat/layout/activity_review.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".UI.ReviewActivity">
+    tools:context=".UI.profileActivities.ReviewActivity">
     <GridView
         android:id="@+id/gridView"
         android:layout_width="0dp"
